[TAS-92] modify social 로그인 성공 후 폴더 이름 및 도메인 이름 변경
