+++ build.gradle
@@ -43,3 +43,14 @@ dependencies {
 tasks.named('test') {
 	useJUnitPlatform()
 }
+// profile의 기본값 local로 설정
+ext.profile = (!project.hasProperty('profile') || !profile) ? 'local' : profile
+
+// 리소스 디렉토리 추가
+sourceSets {
+    main {
+        resources {
+            srcDirs "src/main/resources-env/${profile}"
+        }
+    }
+}
