--- src/main/java/com/aaa/zeppy/OAuth2/handler/OAuth2LoginSuccessHandler.java
@@ -1,78 +1,0 @@
-package com.aaa.zeppy.OAuth2.handler;
-
-import com.aaa.zeppy.user.entity.Role;
-import com.aaa.zeppy.user.entity.User;
-import com.aaa.zeppy.user.repository.UserRepository;
-import com.aaa.zeppy.common.jwt.service.JwtService;
-import com.aaa.zeppy.OAuth2.entity.CustomOAuth2User;
-
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@Slf4j
-@Component
-@RequiredArgsConstructor
-public class OAuth2LoginSuccessHandler implements AuthenticationSuccessHandler {
-
-    private final JwtService jwtService;
-   private final UserRepository userRepository;
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException, ServletException {
-        log.info("OAuth2 Login 성공!");
-        try {
-            CustomOAuth2User oAuth2User = (CustomOAuth2User) authentication.getPrincipal();
-
-            // User의 Role이 GUEST일 경우 처음 요청한 회원이므로 회원가입 페이지로 리다이렉트
-            if(oAuth2User.getRole() == Role.GUEST) {
-                String accessToken = jwtService.createAccessToken(oAuth2User.getEmail());
-                response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
-                response.sendRedirect("com.aaa://"); // 프론트의 회원가입 추가 정보 입력 폼으로 리다이렉트
-                log.info("OAuth2 Login 22222222222222");
-
-                jwtService.sendAccessAndRefreshToken(response, accessToken, null);
-                User findUser = userRepository.findByEmail(oAuth2User.getEmail())
-                               .orElseThrow(() -> new IllegalArgumentException("이메일에 해당하는 유저가 없습니다."));
-
-                log.info("oAuth2User : {}", oAuth2User);
-                log.info("findUser: {}", findUser);
-                findUser.authorizeUser();
-            } else {
-                log.info("OAuth2 Login 3333333333333333");
-                loginSuccess(response, oAuth2User); // 로그인에 성공한 경우 access, refresh 토큰 생성
-            }
-            // response.setStatus(HttpServletResponse.SC_OK);
-            String url = UriComponentsBuilder.fromUriString("com.aaa://")
-                    .queryParam("test", "accc")
-                    .queryParam("test", "resss")
-                    .toUriString();
-            log.info("url: {}", url);
-            log.info("request: {}", request);
-            log.info("OAuth2 Login 4444444444444444444444");
-            log.info("response: {}", response);
-        } catch (Exception e) {
-            log.info("55555555555555555555");
-            throw e;
-        }
-    }
-
-    // TODO : 소셜 로그인 시에도 무조건 토큰 생성하지 말고 JWT 인증 필터처럼 RefreshToken 유/무에 따라 다르게 처리해보기
-    private void loginSuccess(HttpServletResponse response, CustomOAuth2User oAuth2User) throws IOException {
-        String accessToken = jwtService.createAccessToken(oAuth2User.getEmail());
-        String refreshToken = jwtService.createRefreshToken();
-        response.addHeader(jwtService.getAccessHeader(), "Bearer " + accessToken);
-        response.addHeader(jwtService.getRefreshHeader(), "Bearer " + refreshToken);
-
-        jwtService.sendAccessAndRefreshToken(response, accessToken, refreshToken);
-        jwtService.updateRefreshToken(oAuth2User.getEmail(), refreshToken);
-    }
-}
