--- src/main/java/com/aaa/zeppy/OAuth2/userinfo/GoogleOAuth2UserInfo.java
@@ -1,29 +1,0 @@
-package com.aaa.zeppy.OAuth2.userinfo;
-
-import java.util.Map;
-
-public class GoogleOAuth2UserInfo extends OAuth2UserInfo {
-
-    public GoogleOAuth2UserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        return (String) attributes.get("sub");
-    }
-
-    @Override
-    public String getNickname() {
-        return (String) attributes.get("name");
-    }
-
-    @Override
-    public String getImageUrl() {
-        return (String) attributes.get("picture");
-    }
-}
-    // @Override
-    // public String getEmail() {
-    //     return (String) attributes.get("email");
-    // }
