--- src/main/java/com/aaa/zeppy/OAuth2/userinfo/KakaoOAuth2UserInfo.java
@@ -1,39 +1,0 @@
-package com.aaa.zeppy.OAuth2.userinfo;
-
-import java.util.Map;
-
-public class KakaoOAuth2UserInfo extends OAuth2UserInfo {
-
-    public KakaoOAuth2UserInfo(Map<String, Object> attributes) {
-        super(attributes);
-    }
-
-    @Override
-    public String getId() {
-        return String.valueOf(attributes.get("id"));
-    }
-
-    @Override
-    public String getNickname() {
-        Map<String, Object> account = (Map<String, Object>) attributes.get("kakao_account");
-        Map<String, Object> profile = (Map<String, Object>) account.get("profile");
-
-        if (account == null || profile == null) {
-            return null;
-        }
-
-        return (String) profile.get("nickname");
-    }
-
-    @Override
-    public String getImageUrl() {
-        Map<String, Object> account = (Map<String, Object>) attributes.get("kakao_account");
-        Map<String, Object> profile = (Map<String, Object>) account.get("profile");
-
-        if (account == null || profile == null) {
-            return null;
-        }
-
-        return (String) profile.get("thumbnail_image_url");
-    }
-}
