--- src/main/java/com/aaa/zeppy/OAuth2/userinfo/OAuth2UserInfo.java
@@ -1,18 +1,0 @@
-package com.aaa.zeppy.OAuth2.userinfo;
-
-import java.util.Map;
-
-public abstract class OAuth2UserInfo {
-
-    protected Map<String, Object> attributes;
-
-    public OAuth2UserInfo(Map<String, Object> attributes) {
-        this.attributes = attributes;
-    }
-
-    public abstract String getId(); //소셜 식별 값 : 구글 - "sub", 카카오 - "id", 네이버 - "id"
-
-    public abstract String getNickname();
-
-    public abstract String getImageUrl();
-}
