--- src/main/java/com/aaa/zeppy/ZeppyApplication.java
@@ -1,13 +1,0 @@
-package com.aaa.zeppy;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ZeppyApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(ZeppyApplication.class, args);
-	}
-
-}
