--- src/main/java/com/aaa/zeppy/user/dto/UserSignUpDto.java
@@ -1,15 +1,0 @@
-package com.aaa.zeppy.user.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@Getter
-public class UserSignUpDto {
-
-    private String email;
-    private String password;
-    private String nickname;
-    private int age;
-    private String city;
-}
