--- src/main/java/com/aaa/zeppy/user/entity/Role.java
@@ -1,14 +1,0 @@
-package com.aaa.zeppy.user.entity;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum Role {
-    GUEST("ROLE_GUEST"),
-    USER("ROLE_USER"),
-    ADMIN("ROLE_ADMIN");
-
-    private final String key;
-}
