--- src/main/java/com/aaa/zeppy/user/entity/SocialType.java
@@ -1,8 +1,0 @@
-package com.aaa.zeppy.user.entity;
-
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-public enum SocialType {
-    KAKAO, GOOGLE
-}
