--- src/main/java/com/aaa/zeppy/user/entity/User.java
@@ -1,68 +1,0 @@
-package com.aaa.zeppy.user.entity;
-
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Builder
-@Table(name = "USERS")
-@AllArgsConstructor
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "user_id")
-    private Long id;
-
-    private String email; // 이메일
-    private String password; // 비밀번호
-    private String nickname; // 닉네임
-    private String imageUrl; // 프로필 이미지
-    private int age;
-    private String city; // 사는 도시
-
-    @Enumerated(EnumType.STRING)
-    private Role role;
-
-    @Enumerated(EnumType.STRING)
-    private SocialType socialType; // KAKAO, NAVER, GOOGLE
-
-    private String socialId; // 로그인한 소셜 타입의 식별자 값 (일반 로그인인 경우 null)
-
-    private String refreshToken; // 리프레시 토큰
-
-    // 유저 권한 설정 메소드
-    public void authorizeUser() {
-        this.role = Role.USER;
-    }
-
-    // 비밀번호 암호화 메소드
-    public void passwordEncode(PasswordEncoder passwordEncoder) {
-        this.password = passwordEncoder.encode(this.password);
-    }
-
-    //== 유저 필드 업데이트 ==//
-    public void updateNickname(String updateNickname) {
-        this.nickname = updateNickname;
-    }
-
-    public void updateAge(int updateAge) {
-        this.age = updateAge;
-    }
-
-    public void updateCity(String updateCity) {
-        this.city = updateCity;
-    }
-
-    public void updatePassword(String updatePassword, PasswordEncoder passwordEncoder) {
-        this.password = passwordEncoder.encode(updatePassword);
-    }
-
-    public void updateRefreshToken(String updateRefreshToken) {
-        this.refreshToken = updateRefreshToken;
-    }
-}
