+++ src/main/java/kr/co/zeppy/ZeppyApplication.java
@@ -1,0 +1,13 @@
+package kr.co.zeppy;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ZeppyApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ZeppyApplication.class, args);
+	}
+
+}
