+++ src/main/java/kr/co/zeppy/user/controller/UserController.java
@@ -1,0 +1,27 @@
+package kr.co.zeppy.user.controller;
+
+import kr.co.zeppy.user.dto.UserSignUpDto;
+import kr.co.zeppy.user.service.UserService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@RestController
+@RequiredArgsConstructor
+public class UserController {
+
+    private final UserService userService;
+
+    @PostMapping("/sign-up")
+    public String signUp(@RequestBody UserSignUpDto userSignUpDto) throws Exception {
+        userService.signUp(userSignUpDto);
+        return "회원가입 성공";
+    }
+
+    @GetMapping("/jwt-test")
+    public String jwtTest() {
+        return "jwtTest 요청 성공";
+    }
+}
