+++ src/main/java/kr/co/zeppy/user/dto/UserSignUpDto.java
@@ -1,0 +1,15 @@
+package kr.co.zeppy.user.dto;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@NoArgsConstructor
+@Getter
+public class UserSignUpDto {
+
+    private String email;
+    private String password;
+    private String nickname;
+    private int age;
+    private String city;
+}
