+++ src/main/java/kr/co/zeppy/user/entity/Role.java
@@ -1,0 +1,14 @@
+package kr.co.zeppy.user.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public enum Role {
+    GUEST("ROLE_GUEST"),
+    USER("ROLE_USER"),
+    ADMIN("ROLE_ADMIN");
+
+    private final String key;
+}
