+++ src/main/java/kr/co/zeppy/user/entity/SocialType.java
@@ -1,0 +1,8 @@
+package kr.co.zeppy.user.entity;
+
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+public enum SocialType {
+    KAKAO, GOOGLE
+}
