--- src/test/java/com/aaa/zeppy/ZeppyApplicationTests.java
@@ -1,13 +1,0 @@
-package com.aaa.zeppy;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ZeppyApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
