+++ src/test/java/kr/co/zeppy/ZeppyApplicationTests.java
@@ -1,0 +1,13 @@
+package kr.co.zeppy;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ZeppyApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
